Move error reporting into its own source file
This doesn't really seem to belong in save_git.c.
